refactor: ConnectionConsumer jdbc로 이동
