+++ jdbc/src/main/java/com/interface21/jdbc/core/ConnectionConsumer.java
@@ -1,4 +1,4 @@
-package com.techcourse.dao;
+package com.interface21.jdbc.core;
 
 import java.sql.Connection;
 import java.sql.SQLException;
